requirements.txt
asyncpg==0.27.0
fastapi==0.95.1
hsemotion==0.3.0
pillow==9.5.0
psycopg2-binary==2.9.6
pytube==15.0.0
rmn==3.1.1
timm==0.6.5
uvicorn==0.22.0
yoyo-migrations==7.3.2
